db相关的方法移到db
